% Generated by roxygen2 (4.0.1): do not edit by hand
\name{avro_make_schema}
\alias{avro_make_schema}
\title{Create Avro Schema}
\usage{
avro_make_schema(x, name = NULL, namespace = NULL, unflatten = F,
  row.names = T)
}
\arguments{
\item{x}{data object for which to create an Avro schema}

\item{name}{(optional)a "name" value for the top-level Avro schema element}

\item{namespace}{(optional)a "namespace" value for the top-level Avro schema element}

\item{row.names}{should the \code{row.names} attribute be exported as the field \code{row_names}.}
}
\description{
Create an Avro schema object for an R value
}
\details{
Create an Avro schema and convert an R \code{data.frame} value into a list of data elements.

Optional \code{name} and \code{namespace} arguments can be used to specify the top-level Avro
element \code{name} and \code{namespace}.

If \code{is.union = TRUE}, each data element will be wrapped in a \code{list} named with the
Avro type for that data element, consistent with the Avro JSON encoding.

If \code{is.field = TRUE}, slight modifications will be made to the Avro schema and data
elements, allowing them to be used as values/schemas for a Avro \code{record}'s "field" value.
}

