% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{package}
\name{ravro-package}
\alias{ravro-package}
\title{Read and write Avro data files from R.}
\description{
The ravro package provides tools to allow reading of Avro data files into R.
It provides the functions \code{\link{read.avro}}
for reading Avro data files.
}
\section{Dependencies}{

\subsection{R Packages}{
Prior to installing the ravro package, the following R packages must be
installed: Rcpp, rjson, bit64 and bit.  These can be installed in R with
the command:

\code{install.packages(c("Rcpp","rjson","bit64"))}

This command will install these packages, as well as the "bit" package,
which is required by bit64.
}

\subsection{Java}{
In addition, Java must be installed on the system and be available via
the \code{PATH} environmental variable. Try running the following from R:

\code{system("java -version")}

The Java version must be at least 1.6
}
}

\section{Installation}{

After these dependencies have been installed, ravro can be installed
from the command line:

\code{R CMD INSTALL ravro_1.0.zip}

Or through an R development environment, e.g. RStudio.

You can verify that the package is working correctly on your system with

\code{R CMD check ravro_1.0.zip}

NOTE: On Linux and Mac OS systems, you will install and/or check the source
package: \code{ravro_1.0.tar.gz}.
}
\references{
Apache Avro 1.7.6 Specification. \url{http://avro.apache.org/docs/1.7.6/spec.html}.
}

